requirements.txt
mysql-connector-python==8.0.22
protobuf==3.14.0
PyYAML==5.3.1
six==1.15.0
